move helpers for testing events to EventTesterUtils
